backend/src/requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
certifi==2019.3.9
chardet==3.0.4
constantly==15.1.0
daphne==2.2.5
Django==2.1.7
gunicorn
uvicorn
channels-redis==2.3.1
channels==2.1.5
django-braces==1.13.0
django-cors-headers==2.5.1
djangorestframework==3.9.2
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
oauthlib==3.0.1
PyHamcrest==1.9.0
pytz==2018.9
psycopg2-binary
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
zope.interface==4.6.0